+++ src/main/kotlin/com/github/noamm9/features/impl/dungeon/Secrets.kt
@@ -1,23 +1,14 @@
 package com.github.noamm9.features.impl.dungeon
 
-import com.github.noamm9.config.types.ColorSetting
-import com.github.noamm9.config.types.DropdownSetting
-import com.github.noamm9.config.types.SliderSetting
-import com.github.noamm9.config.types.ToggleSetting
-import com.github.noamm9.event.impl.DungeonEvent
-import com.github.noamm9.event.impl.GameStartEvent
-import com.github.noamm9.event.impl.MainThreadPacketReceivedEvent
-import com.github.noamm9.event.impl.RenderWorldEvent
+import com.github.noamm9.config.types.*
+import com.github.noamm9.event.impl.*
 import com.github.noamm9.features.Feature
 import com.github.noamm9.init.ModCompatibility
-import com.github.noamm9.utils.ActionBarParser
+import com.github.noamm9.utils.*
 import com.github.noamm9.utils.ChatUtils.unformattedText
-import com.github.noamm9.utils.ColorUtils
 import com.github.noamm9.utils.ColorUtils.withAlpha
-import com.github.noamm9.utils.Utils
 import com.github.noamm9.utils.Utils.send
 import com.github.noamm9.utils.dungeons.enums.SecretType
-import com.github.noamm9.utils.equalsOneOf
 import com.github.noamm9.utils.location.LocationUtils
 import com.github.noamm9.utils.render.Render2D.drawString
 import com.github.noamm9.utils.render.RenderHelper.width
@@ -27,9 +18,7 @@ import net.minecraft.network.protocol.game.ClientboundOpenScreenPacket
 import net.minecraft.network.protocol.game.ServerboundContainerClosePacket
 import net.minecraft.sounds.SoundEvents
 import net.minecraft.world.inventory.MenuType
-import net.minecraft.world.level.block.Block
-import net.minecraft.world.level.block.ButtonBlock
-import net.minecraft.world.level.block.FaceAttachedHorizontalDirectionalBlock
+import net.minecraft.world.level.block.*
 import net.minecraft.world.level.block.state.BlockState
 import net.minecraft.world.phys.shapes.Shapes
 import net.minecraft.world.phys.shapes.VoxelShape
@@ -70,7 +59,7 @@ object Secrets: Feature() {
             }
 
             ctx.drawString(line, 0, 0)
-            return@hudElement line.width().toFloat() to 9f
+            return@hudElement line.width() to 9f
         }
 
         register<RenderWorldEvent> {
